php: building an docEditorView from a templates + formatting by linter
